needed to fix the tests for json.Number

+++ client_test.go
@@ -437,7 +437,7 @@ func TestClient_Select(t *testing.T) {
 	})
 
 	exp := [][]interface{}{
-		[]interface{}{"1970-01-01T00:00:00Z", float64(5)},
+		[]interface{}{"1970-01-01T00:00:00Z", int64(5)},
 	}
 	if !reflect.DeepEqual(got, exp) {
 		t.Fatalf("Values = %q; want %q", got, exp)
